Move JavaPackageFinder (and :packagefinder) up into jvm.core
Summary:
While implementations of the JavaPackageFinder interface may be
language specific, the interface is not, so move it to core.

Test plan: CI
