+++ src/com/facebook/buck/cli/BUCK
@@ -6,9 +6,9 @@ java_library(
   name = 'events',
   srcs = EVENTS_SRCS,
   deps = [
-    '//third-party/java/guava:guava',
     '//src/com/facebook/buck/event:event',
     '//src/com/facebook/buck/model:model',
+    '//third-party/java/guava:guava',
   ],
   visibility = [
     'PUBLIC',
@@ -46,6 +46,7 @@ java_library(
   deps =  [
     '//src/com/facebook/buck/event:event',
     '//src/com/facebook/buck/io:io',
+    '//src/com/facebook/buck/jvm/core:packagefinder',
     '//src/com/facebook/buck/jvm/java:packagefinder',
     '//src/com/facebook/buck/jvm/java:support',
     '//src/com/facebook/buck/log:api',
@@ -111,9 +112,9 @@ java_immutables_library(
     '//src/com/facebook/buck/android:packageable',
     '//src/com/facebook/buck/android:rules',
     '//src/com/facebook/buck/android:utils',
+    '//src/com/facebook/buck/apple:rules',
     '//src/com/facebook/buck/apple/device:device',
     '//src/com/facebook/buck/apple/simulator:simulator',
-    '//src/com/facebook/buck/apple:rules',
     '//src/com/facebook/buck/artifact_cache:artifact_cache',
     '//src/com/facebook/buck/artifact_cache:config',
     '//src/com/facebook/buck/command:command',
@@ -124,10 +125,13 @@ java_immutables_library(
     '//src/com/facebook/buck/file:downloader',
     '//src/com/facebook/buck/file:file',
     '//src/com/facebook/buck/graph:graph',
+    '//src/com/facebook/buck/httpserver:httpserver',
     '//src/com/facebook/buck/io:io',
     '//src/com/facebook/buck/io:watchman',
     '//src/com/facebook/buck/js:js',
+    '//src/com/facebook/buck/json:json',
     '//src/com/facebook/buck/jvm/core:classhash',
+    '//src/com/facebook/buck/jvm/core:packagefinder',
     '//src/com/facebook/buck/jvm/java:autodeps',
     '//src/com/facebook/buck/jvm/java:config',
     '//src/com/facebook/buck/jvm/java:packagefinder',
@@ -135,7 +139,6 @@ java_immutables_library(
     '//src/com/facebook/buck/jvm/java:steps',
     '//src/com/facebook/buck/jvm/java:support',
     '//src/com/facebook/buck/jvm/java/intellij:intellij',
-    '//src/com/facebook/buck/json:json',
     '//src/com/facebook/buck/log:api',
     '//src/com/facebook/buck/log:log',
     '//src/com/facebook/buck/maven:maven',
@@ -149,35 +152,34 @@ java_immutables_library(
     '//src/com/facebook/buck/rules:build_rule',
     '//src/com/facebook/buck/rules:rules',
     '//src/com/facebook/buck/rules:target_graph_to_action_graph',
-    '//src/com/facebook/buck/rules/keys:keys',
-    '//src/com/facebook/buck/test:report-format',
-    '//src/com/facebook/buck/test:config',
-    '//src/com/facebook/buck/test:test',
-    '//src/com/facebook/buck/test/result/groups:groups',
     '//src/com/facebook/buck/rules:types',
-    '//src/com/facebook/buck/httpserver:httpserver',
+    '//src/com/facebook/buck/rules/keys:keys',
     '//src/com/facebook/buck/shell:steps',
     '//src/com/facebook/buck/simulate:simulate',
     '//src/com/facebook/buck/step:step',
     '//src/com/facebook/buck/step/fs:fs',
+    '//src/com/facebook/buck/test:config',
+    '//src/com/facebook/buck/test:report-format',
+    '//src/com/facebook/buck/test:test',
+    '//src/com/facebook/buck/test/result/groups:groups',
     '//src/com/facebook/buck/test/result/type:type',
     '//src/com/facebook/buck/test/selectors:selectors',
+    '//src/com/facebook/buck/timing:timing',
     '//src/com/facebook/buck/util:constants',
     '//src/com/facebook/buck/util:escaper',
     '//src/com/facebook/buck/util:exceptions',
     '//src/com/facebook/buck/util:io',
     '//src/com/facebook/buck/util:util',
     '//src/com/facebook/buck/util:watchman',
-    '//src/com/facebook/buck/util/unit:unit',
     '//src/com/facebook/buck/util/cache:cache',
     '//src/com/facebook/buck/util/concurrent:concurrent',
-    '//src/com/facebook/buck/util/environment:environment',
     '//src/com/facebook/buck/util/environment:env-filter',
+    '//src/com/facebook/buck/util/environment:environment',
     '//src/com/facebook/buck/util/environment:platform',
     '//src/com/facebook/buck/util/network:network',
-    '//src/com/facebook/buck/timing:timing',
-    '//src/com/facebook/buck/util/versioncontrol:versioncontrol',
+    '//src/com/facebook/buck/util/unit:unit',
     '//src/com/facebook/buck/util/versioncontrol:impl',
+    '//src/com/facebook/buck/util/versioncontrol:versioncontrol',
     '//third-party/java/aether:aether-api',
     '//third-party/java/args4j:args4j',
     '//third-party/java/ddmlib:ddmlib',
