+++ src/com/facebook/buck/command/BUCK
@@ -15,7 +15,7 @@ java_immutables_library(
     '//src/com/facebook/buck/graph:graph',
     '//src/com/facebook/buck/io:io',
     '//src/com/facebook/buck/jvm/core:classhash',
-    '//src/com/facebook/buck/jvm/java:packagefinder',
+    '//src/com/facebook/buck/jvm/core:packagefinder',
     '//src/com/facebook/buck/jvm/java:rules',
     '//src/com/facebook/buck/jvm/java:support',
     '//src/com/facebook/buck/model:model',
