+++ src/com/facebook/buck/jvm/java/BUCK
@@ -21,11 +21,11 @@ java_library(
   ],
   deps = [
     ':support',
-    '//third-party/java/guava:guava',
     '//src/com/facebook/buck/cli:config',
     '//src/com/facebook/buck/rules:build_rule',
     '//src/com/facebook/buck/util:exceptions',
     '//src/com/facebook/buck/util:io',
+    '//third-party/java/guava:guava',
   ],
   visibility = ['PUBLIC'],
 )
@@ -70,10 +70,10 @@ java_immutables_library(
   ],
   deps = [
     ':javac-internals',
-    ':packagefinder',
     '//src/com/facebook/buck/event:event',
     '//src/com/facebook/buck/io:io',
     '//src/com/facebook/buck/jvm/core:classhash',
+    '//src/com/facebook/buck/jvm/core:packagefinder',
     '//src/com/facebook/buck/jvm/java/abi:protocol',
     '//src/com/facebook/buck/jvm/java/tracing:api',
     '//src/com/facebook/buck/jvm/java/tracing:tracing',
@@ -102,11 +102,11 @@ java_library(
   name = 'packagefinder',
   srcs = [
     'DefaultJavaPackageFinder.java',
-    'JavaPackageFinder.java',
     'ResourcesRootPackageFinder.java',
   ],
   deps = [
     '//src/com/facebook/buck/io:io',
+    '//src/com/facebook/buck/jvm/core:packagefinder',
     '//src/com/facebook/buck/model:model',
     '//src/com/facebook/buck/util:exceptions',
     '//third-party/java/guava:guava',
@@ -160,6 +160,7 @@ java_library(
     '//src/com/facebook/buck/io:io',
     '//src/com/facebook/buck/jvm/core:classhash',
     '//src/com/facebook/buck/jvm/core:nativelinkable',
+    '//src/com/facebook/buck/jvm/core:packagefinder',
     '//src/com/facebook/buck/jvm/java/abi:protocol',
     '//src/com/facebook/buck/log:api',
     '//src/com/facebook/buck/maven:util',
@@ -168,16 +169,16 @@ java_library(
     '//src/com/facebook/buck/rules:build_rule',
     '//src/com/facebook/buck/rules:rules',
     '//src/com/facebook/buck/rules/keys:keys',
-    '//src/com/facebook/buck/test:test',
     '//src/com/facebook/buck/shell:steps',
     '//src/com/facebook/buck/step:step',
     '//src/com/facebook/buck/step/fs:fs',
+    '//src/com/facebook/buck/test:test',
+    '//src/com/facebook/buck/test/result/type:type',
     '//src/com/facebook/buck/test/selectors:selectors',
     '//src/com/facebook/buck/util:constants',
     '//src/com/facebook/buck/util:exceptions',
     '//src/com/facebook/buck/util:util',
     '//src/com/facebook/buck/zip:steps',
-    '//src/com/facebook/buck/test/result/type:type',
     '//third-party/java/guava:guava',
     '//third-party/java/infer-annotations:infer-annotations',
     '//third-party/java/jackson:jackson',
@@ -208,11 +209,12 @@ java_library(
     ':packagefinder',
     ':support',
     '//src/com/facebook/buck/dalvik:dalvik_stats_tool',
+    '//src/com/facebook/buck/event:event',
+    '//src/com/facebook/buck/io:io',
+    '//src/com/facebook/buck/jvm/core:packagefinder',
     '//src/com/facebook/buck/jvm/java/abi:abi',
     '//src/com/facebook/buck/jvm/java/classes:classes',
     '//src/com/facebook/buck/jvm/java/runner:runner',
-    '//src/com/facebook/buck/event:event',
-    '//src/com/facebook/buck/io:io',
     '//src/com/facebook/buck/log:api',
     '//src/com/facebook/buck/model:model',
     '//src/com/facebook/buck/rules:build_rule',
@@ -252,6 +254,7 @@ java_library(
     '//src/com/facebook/buck/event:event',
     '//src/com/facebook/buck/io:io',
     '//src/com/facebook/buck/json:json',
+    '//src/com/facebook/buck/jvm/core:packagefinder',
     '//src/com/facebook/buck/model:model',
     '//src/com/facebook/buck/parser:config',
     '//src/com/facebook/buck/rules:build_rule',
