+++ src/com/facebook/buck/jvm/java/CopyResourcesStep.java
@@ -17,6 +17,7 @@ package com.facebook.buck.jvm.java;
 
 import com.facebook.buck.io.MorePaths;
 import com.facebook.buck.io.ProjectFilesystem;
+import com.facebook.buck.jvm.core.JavaPackageFinder;
 import com.facebook.buck.model.BuildTarget;
 import com.facebook.buck.model.HasOutputName;
 import com.facebook.buck.rules.BuildRule;
