+++ src/com/facebook/buck/jvm/java/DefaultJavaLibrary.java
@@ -25,6 +25,7 @@ import com.facebook.buck.android.AndroidPackageableCollector;
 import com.facebook.buck.graph.DirectedAcyclicGraph;
 import com.facebook.buck.graph.TopologicalSort;
 import com.facebook.buck.io.ProjectFilesystem;
+import com.facebook.buck.jvm.core.JavaPackageFinder;
 import com.facebook.buck.model.BuildTarget;
 import com.facebook.buck.model.BuildTargets;
 import com.facebook.buck.model.HasTests;
