+++ src/com/facebook/buck/jvm/java/JavaSourceJar.java
@@ -18,6 +18,7 @@ package com.facebook.buck.jvm.java;
 
 import static com.facebook.buck.zip.ZipStep.DEFAULT_COMPRESSION_LEVEL;
 
+import com.facebook.buck.jvm.core.JavaPackageFinder;
 import com.facebook.buck.model.BuildTarget;
 import com.facebook.buck.model.BuildTargets;
 import com.facebook.buck.rules.AbstractBuildRule;
