+++ src/com/facebook/buck/jvm/java/JavacErrorParser.java
@@ -17,6 +17,7 @@
 package com.facebook.buck.jvm.java;
 
 import com.facebook.buck.io.ProjectFilesystem;
+import com.facebook.buck.jvm.core.JavaPackageFinder;
 import com.google.common.base.Optional;
 import com.google.common.collect.ImmutableList;
