+++ src/com/facebook/buck/jvm/java/intellij/IjSourceRootSimplifier.java
@@ -18,7 +18,7 @@ package com.facebook.buck.jvm.java.intellij;
 
 import com.facebook.buck.graph.MutableDirectedGraph;
 import com.facebook.buck.io.MorePaths;
-import com.facebook.buck.jvm.java.JavaPackageFinder;
+import com.facebook.buck.jvm.core.JavaPackageFinder;
 import com.facebook.buck.util.immutables.BuckStyleImmutable;
 import com.google.common.base.Function;
 import com.google.common.base.Optional;
