+++ src/com/facebook/buck/jvm/java/intellij/ParsingJavaPackageFinder.java
@@ -18,9 +18,9 @@ package com.facebook.buck.jvm.java.intellij;
 
 import com.facebook.buck.io.MorePaths;
 import com.facebook.buck.io.ProjectFilesystem;
+import com.facebook.buck.jvm.core.JavaPackageFinder;
 import com.facebook.buck.jvm.java.DefaultJavaPackageFinder;
 import com.facebook.buck.jvm.java.JavaFileParser;
-import com.facebook.buck.jvm.java.JavaPackageFinder;
 import com.facebook.buck.model.BuildTarget;
 import com.facebook.buck.util.Optionals;
 import com.google.common.base.Function;
