+++ src/com/facebook/buck/jvm/java/intellij/Project.java
@@ -34,9 +34,9 @@ import com.facebook.buck.android.NdkLibrary;
 import com.facebook.buck.cxx.CxxLibrary;
 import com.facebook.buck.graph.AbstractBreadthFirstTraversal;
 import com.facebook.buck.io.ProjectFilesystem;
+import com.facebook.buck.jvm.core.JavaPackageFinder;
 import com.facebook.buck.jvm.java.JavaBinary;
 import com.facebook.buck.jvm.java.JavaLibrary;
-import com.facebook.buck.jvm.java.JavaPackageFinder;
 import com.facebook.buck.jvm.java.PrebuiltJar;
 import com.facebook.buck.log.Logger;
 import com.facebook.buck.model.BuildFileTree;
