+++ src/com/facebook/buck/rules/BUCK
@@ -17,6 +17,7 @@ java_library(
     '//src/com/facebook/buck/apple:rules',
     '//src/com/facebook/buck/cli:config',
     '//src/com/facebook/buck/cxx:platform',
+    '//src/com/facebook/buck/cxx:rules',
     '//src/com/facebook/buck/d:d',
     '//src/com/facebook/buck/dotnet:dotnet',
     '//src/com/facebook/buck/event:event',
@@ -27,26 +28,25 @@ java_library(
     '//src/com/facebook/buck/halide:rules',
     '//src/com/facebook/buck/io:io',
     '//src/com/facebook/buck/io:watchman',
+    '//src/com/facebook/buck/js:js',
+    '//src/com/facebook/buck/json:json',
     '//src/com/facebook/buck/jvm/java:config',
     '//src/com/facebook/buck/jvm/java:rules',
     '//src/com/facebook/buck/jvm/java:support',
-    '//src/com/facebook/buck/js:js',
-    '//src/com/facebook/buck/json:json',
     '//src/com/facebook/buck/log:api',
     '//src/com/facebook/buck/log:log',
     '//src/com/facebook/buck/model:model',
+    '//src/com/facebook/buck/ocaml:rules',
     '//src/com/facebook/buck/parser:config',
     '//src/com/facebook/buck/parser:rule_pattern',
     '//src/com/facebook/buck/python:config',
-    '//src/com/facebook/buck/python:support',
     '//src/com/facebook/buck/python:rules',
-    '//src/com/facebook/buck/ocaml:rules',
-    '//src/com/facebook/buck/cxx:rules',
+    '//src/com/facebook/buck/python:support',
     '//src/com/facebook/buck/rules/macros:macros',
     '//src/com/facebook/buck/rust:rust',
     '//src/com/facebook/buck/shell:rules',
-    '//src/com/facebook/buck/timing:timing',
     '//src/com/facebook/buck/thrift:rules',
+    '//src/com/facebook/buck/timing:timing',
     '//src/com/facebook/buck/util:constants',
     '//src/com/facebook/buck/util:exceptions',
     '//src/com/facebook/buck/util:io',
@@ -149,15 +149,15 @@ java_immutables_library(
   ],
   deps = [
     ':rule_key',
-    '//src/com/facebook/buck/apple/xcode:xcode',
     '//src/com/facebook/buck/android:exceptions',
     '//src/com/facebook/buck/android:utils',
+    '//src/com/facebook/buck/apple/xcode:xcode',
     '//src/com/facebook/buck/artifact_cache:artifact_cache',
     '//src/com/facebook/buck/event:event',
     '//src/com/facebook/buck/graph:graph',
     '//src/com/facebook/buck/hashing:hashing',
-    '//src/com/facebook/buck/jvm/java:packagefinder',
     '//src/com/facebook/buck/io:io',
+    '//src/com/facebook/buck/jvm/core:packagefinder',
     '//src/com/facebook/buck/log:api',
     '//src/com/facebook/buck/model:model',
     '//src/com/facebook/buck/parser:rule_pattern',
@@ -245,8 +245,8 @@ java_immutables_library(
     '//test/com/facebook/buck/rules:rules',
   ],
   deps = [
-    ':build_rule',
     ':build_event',
+    ':build_rule',
     '//src/com/facebook/buck/android:exceptions',
     '//src/com/facebook/buck/artifact_cache:artifact_cache',
     '//src/com/facebook/buck/cli:events',
@@ -268,15 +268,15 @@ java_immutables_library(
     '//src/com/facebook/buck/util:util',
     '//src/com/facebook/buck/util/cache:cache',
     '//src/com/facebook/buck/util/collect:collect',
-    '//src/com/facebook/buck/util/hash:hash',
     '//src/com/facebook/buck/util/concurrent:concurrent',
     '//src/com/facebook/buck/util/environment:environment',
+    '//src/com/facebook/buck/util/hash:hash',
     '//src/com/facebook/buck/util/network:network',
     '//src/com/facebook/buck/zip:steps',
     '//src/com/facebook/buck/zip:unzip',
     '//third-party/java/commons-compress:commons-compress',
-    '//third-party/java/infer-annotations:infer-annotations',
     '//third-party/java/guava:guava',
+    '//third-party/java/infer-annotations:infer-annotations',
     '//third-party/java/jackson:jackson-annotations',
     '//third-party/java/jackson:jackson-core',
     '//third-party/java/jackson:jackson-databind',
