+++ src/com/facebook/buck/step/BUCK
@@ -9,17 +9,17 @@ java_immutables_library(
     '//src/com/facebook/buck/android:utils',
     '//src/com/facebook/buck/event:event',
     '//src/com/facebook/buck/io:io',
-    '//src/com/facebook/buck/jvm/java:packagefinder',
+    '//src/com/facebook/buck/jvm/core:packagefinder',
     '//src/com/facebook/buck/log:api',
     '//src/com/facebook/buck/model:model',
     '//src/com/facebook/buck/util:exceptions',
     '//src/com/facebook/buck/util:io',
     '//src/com/facebook/buck/util:util',
-    '//src/com/facebook/buck/util/environment:platform',
     '//src/com/facebook/buck/util/concurrent:concurrent',
+    '//src/com/facebook/buck/util/environment:platform',
     '//third-party/java/guava:guava',
-    '//third-party/java/jsr:jsr305',
     '//third-party/java/jackson:jackson',
+    '//third-party/java/jsr:jsr305',
   ],
   visibility = [ 'PUBLIC' ],
 )
