+++ test/com/facebook/buck/android/BUCK
@@ -37,14 +37,14 @@ java_library(
     '//src/com/facebook/buck/jvm/java:rules',
     '//src/com/facebook/buck/jvm/java:support',
     '//src/com/facebook/buck/model:model',
-    '//src/com/facebook/buck/rules/macros:macros',
     '//src/com/facebook/buck/rules:build_rule',
     '//src/com/facebook/buck/rules:rules',
+    '//src/com/facebook/buck/rules/macros:macros',
     '//src/com/facebook/buck/shell:rules',
     '//src/com/facebook/buck/shell:steps',
     '//src/com/facebook/buck/step:step',
-    '//src/com/facebook/buck/util/environment:platform',
     '//src/com/facebook/buck/util:util',
+    '//src/com/facebook/buck/util/environment:platform',
     '//test/com/facebook/buck/cli:FakeBuckConfig',
     '//test/com/facebook/buck/cxx:testutil',
     '//test/com/facebook/buck/jvm/java:testutil',
@@ -52,8 +52,8 @@ java_library(
     '//test/com/facebook/buck/testutil:testutil',
     '//third-party/java/ddmlib:ddmlib',
     '//third-party/java/guava:guava',
-    '//third-party/java/junit:junit',
     '//third-party/java/jsr:jsr305',
+    '//third-party/java/junit:junit',
   ],
   visibility = [
     'PUBLIC',
@@ -86,7 +86,7 @@ java_test(
     '//src/com/facebook/buck/graph:graph',
     '//src/com/facebook/buck/io:io',
     '//src/com/facebook/buck/jvm/core:classhash',
-    '//src/com/facebook/buck/jvm/java:packagefinder',
+    '//src/com/facebook/buck/jvm/core:packagefinder',
     '//src/com/facebook/buck/jvm/java:rules',
     '//src/com/facebook/buck/jvm/java:steps',
     '//src/com/facebook/buck/jvm/java:support',
