+++ test/com/facebook/buck/jvm/java/BUCK
@@ -20,8 +20,8 @@ java_library(
     '//src/com/facebook/buck/io:io',
     '//src/com/facebook/buck/jvm/core:classhash',
     '//src/com/facebook/buck/jvm/java:rules',
-    '//src/com/facebook/buck/jvm/java:support',
     '//src/com/facebook/buck/jvm/java:steps',
+    '//src/com/facebook/buck/jvm/java:support',
     '//src/com/facebook/buck/model:model',
     '//src/com/facebook/buck/parser:rule_pattern',
     '//src/com/facebook/buck/rules:build_rule',
@@ -48,7 +48,9 @@ java_library(
     '//src/com/facebook/buck/model:model',
   ],
   exported_deps = [
+    # this is a bit dubious, exporting the real one with the fake...
     '//src/com/facebook/buck/jvm/java:packagefinder',
+    '//src/com/facebook/buck/jvm/core:packagefinder',
   ],
   visibility = [
     '//test/...',
@@ -102,8 +104,9 @@ java_test(
     '//src/com/facebook/buck/event:event',
     '//src/com/facebook/buck/graph:graph',
     '//src/com/facebook/buck/io:io',
-    '//src/com/facebook/buck/jvm/java:autodeps',
     '//src/com/facebook/buck/jvm/core:classhash',
+    '//src/com/facebook/buck/jvm/core:packagefinder',
+    '//src/com/facebook/buck/jvm/java:autodeps',
     '//src/com/facebook/buck/jvm/java:config',
     '//src/com/facebook/buck/jvm/java:packagefinder',
     '//src/com/facebook/buck/jvm/java:rules',
