+++ test/com/facebook/buck/jvm/java/FakeJavaPackageFinder.java
@@ -16,6 +16,7 @@
 
 package com.facebook.buck.jvm.java;
 
+import com.facebook.buck.jvm.core.JavaPackageFinder;
 import com.facebook.buck.model.BuildTarget;
 
 import java.nio.file.Path;
