+++ test/com/facebook/buck/jvm/java/JavacErrorParserTest.java
@@ -19,6 +19,7 @@ package com.facebook.buck.jvm.java;
 import static org.junit.Assert.assertEquals;
 
 import com.facebook.buck.io.ProjectFilesystem;
+import com.facebook.buck.jvm.core.JavaPackageFinder;
 import com.google.common.base.Optional;
 import com.google.common.collect.ImmutableSet;
