+++ test/com/facebook/buck/jvm/java/intellij/BUCK
@@ -7,11 +7,12 @@ java_test(
     '//src/com/facebook/buck/cli:config',
     '//src/com/facebook/buck/graph:graph',
     '//src/com/facebook/buck/io:io',
-    '//src/com/facebook/buck/jvm/java/intellij:intellij',
+    '//src/com/facebook/buck/jvm/core:packagefinder',
     '//src/com/facebook/buck/jvm/java:autodeps',
     '//src/com/facebook/buck/jvm/java:packagefinder',
     '//src/com/facebook/buck/jvm/java:rules',
     '//src/com/facebook/buck/jvm/java:support',
+    '//src/com/facebook/buck/jvm/java/intellij:intellij',
     '//src/com/facebook/buck/model:model',
     '//src/com/facebook/buck/rules:build_rule',
     '//src/com/facebook/buck/rules:rules',
@@ -27,8 +28,8 @@ java_test(
     '//test/com/facebook/buck/model:BuildTargetFactory',
     '//test/com/facebook/buck/rules:testutil',
     '//test/com/facebook/buck/shell:testutil',
-    '//test/com/facebook/buck/testutil/integration:integration',
     '//test/com/facebook/buck/testutil:testutil',
+    '//test/com/facebook/buck/testutil/integration:integration',
     '//test/com/facebook/buck/timing:testutil',
     '//third-party/java/easymock:easymock',
     '//third-party/java/guava:guava',
