+++ test/com/facebook/buck/jvm/java/intellij/ParsingJavaPackageFinderTest.java
@@ -18,9 +18,9 @@ package com.facebook.buck.jvm.java.intellij;
 
 import static org.junit.Assert.assertEquals;
 
+import com.facebook.buck.jvm.core.JavaPackageFinder;
 import com.facebook.buck.jvm.java.JavaCompilationConstants;
 import com.facebook.buck.jvm.java.JavaFileParser;
-import com.facebook.buck.jvm.java.JavaPackageFinder;
 import com.facebook.buck.model.BuildTarget;
 import com.facebook.buck.testutil.FakeProjectFilesystem;
 import com.facebook.buck.timing.FakeClock;
